Cannot type '我'

+++ include/internal/dict-private.h
@@ -21,7 +21,7 @@
 
 #define PHONE_PHRASE_NUM (162244)
 
-int GetCharFirst(ChewingData *, Phrase *, uint16_t);
+int GetCharFirst(ChewingData *, Phrase *, uint32_t);
 int GetPhraseFirst(ChewingData *pgdata, Phrase *phr_ptr, const TreeType *phrase_parent);
 int GetVocabNext(ChewingData *pgdata, Phrase *phr_ptr);
 int InitDict(ChewingData *pgdata, const char *prefix);
